Moved botocore and warnings imports to the relevant method.
In the case of a configuration that doesn't use AWS, it won't be necessary to have boto3 installed. Doing this allows this module to be used without requiring boto3 unless used.

In the future, it will probably be better to move the AWS hook (and other future hooks) to separate modules.