Created compose.yaml for Docker Compose
Run with `docker compose up` (may need `docker-compose up` instead).

This seems to allow the container to access the internet (i.e. AWS).

Since the compose file builds the dockerfile everytime, it re-fetches the config.json. This is works but may not be the best way to go about it.